requirements.txt
asgiref==3.2.10
astroid==2.4.1
autopep8==1.5
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.1
cycler==0.10.0
Django==2.2.3
django-cors-headers==3.5.0
djangorestframework==3.11.1
idna==2.8
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.5.2
mccabe==0.6.1
numpy==1.16.4
pandas==1.0.1
Pillow==7.2.0
pycodestyle==2.5.0
pylint==2.5.2
pyparsing==2.4.1.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
scipy==1.4.1
selenium==3.141.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.3
virtualenv==16.5.0
webdriver-manager==3.2.1
wrapt==1.11.2
xlrd==1.2.0
